LayoutTests/http/tests/security/script-no-crossorigin-error-event-should-be-sanitized.html

<body>
<script src="../../js-test-resources/js-test-pre.js"></script>
<script>
window.jsTestIsAsync = true;
description("The test passes if error event gets sanitized information about the script error.");

window.addEventListener('error', function() {
    shouldBeNull("event.message.match(/SomeError/)");
    shouldBeEqualToString("event.filename", "");
    shouldBe("event.lineno", "0");
    shouldBe("event.colno", "0");
    shouldBeNull("event.error");
    finishJSTest();
});

successfullyParsed = true;
</script>
<!-- crossorigin attribute is not set, we shouldn't get any specific info on the error. -->
<script src="http://localhost:8000/security/resources/cors-script.php?fail=true"></script>
<script src="../../js-test-resources/js-test-post.js"></script>
