LayoutTests/http/tests/security/script-no-crossorigin-error-event-should-be-sanitized-expected.txt

CONSOLE MESSAGE: line 1: SomeError
The test passes if error event gets sanitized information about the script error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.message.match(/SomeError/) is null
PASS event.filename is ""
PASS event.lineno is 0
PASS event.colno is 0
PASS event.error is null
PASS successfullyParsed is true

TEST COMPLETE
